core: use effect handler for item activation
