+++ LostDreams/GuiltFragmentBonus/GuiltFragmentPatches.cs
@@ -9,7 +9,7 @@ class GuiltDrop_GiveBonus_Patch
 {
     public static void Postfix()
     {
-        if (Main.LostDreams.IsActive("guilt-fragment"))
+        if (Main.LostDreams.EffectHandler.IsActive("guilt-fragment"))
         {
             Main.LostDreams.Log("QI502: Increasing guilt drop bonus");
             Core.Logic.Penitent.Stats.Life.SetToCurrentMax();
